Refactoring Provider and Middleware
